refactor: Change files structure for Training Pages
