+++ src/components/TrainingPages/CreateTrainingPage/CreateTraining.module.scss
