+++ src/components/TrainingPages/CreateTrainingPage/index.ts
