+++ src/components/TrainingPages/Subheader/index.ts
